+++ src/main/java/controller/GameController.java
@@ -2,19 +2,19 @@ package controller;
 
 import domain.Command;
 import domain.GameManager;
-import domain.GameResultDto;
 import domain.Referee;
 import domain.betting.BettingAmount;
 import domain.betting.BettingAmountParser;
 import domain.betting.BettingTable;
 import domain.betting.ParticipantsProfitResult;
 import domain.betting.PlayerBetting;
+import domain.betting.PlayerBettingResult;
 import domain.card.CardsSnapshot;
+import domain.dto.GameResultDto;
 import domain.dto.InitialDealingResult;
 import domain.participant.Dealer;
 import domain.participant.Participants;
 import domain.participant.Player;
-import domain.participant.PlayerBettingResult;
 import domain.participant.PlayerParser;
 import domain.participant.Players;
 import java.util.LinkedHashMap;
