+++ src/main/java/domain/dto/GameResultDto.java
@@ -1,10 +1,11 @@
-package domain;
+package domain.dto;
 
+import domain.MatchResult;
 import java.util.EnumMap;
 import java.util.Map;
 
 public record GameResultDto(
         EnumMap<MatchResult, Integer> dealerMatchResult,
-        Map<String ,MatchResult> playerMatchResults
+        Map<String, MatchResult> playerMatchResults
 ) {
 }
